+++ src/Common/NetworkCore/Network/Rest/Data/ListPlayersResponse.cs
@@ -26,5 +26,5 @@ public class ListPlayersResponse
     /// <summary>
     ///     Players associated with the currently logged in account.
     /// </summary>
-    public IEnumerable<PlayerInfo> Players { get; set; }
+    public IList<PlayerInfo> Players { get; set; }
 }